backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
rdflib==7.0.0
google-generativeai==0.3.2
python-multipart==0.0.6
aiohttp==3.9.5
prometheus_client==0.20.0
jsonschema==4.23.0
aiofiles==23.2.1
PyYAML==6.0.2
watchdog==4.0.1
redis==5.0.1
